To implement Pepper in Native Client we need access to the implementation of the Pepper APIs in Chrome.  Specifically, we need to be able to get the base::SharedMemory or TransportDIB memory regions used to communicate with devices.  To enable doing this, especially in 64-bit Chrome, I needed to change the GetStateContext and SetStateContext functions to manipulate 64-bit values.  This CL does not include the Native Client hookup to the new APIs, but is needed to do so. BUG=none TEST=none
Review URL: http://codereview.chromium.org/569004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@37975 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/renderer/webplugin_delegate_pepper.h
@@ -83,10 +83,10 @@ class WebPluginDelegatePepper : public webkit_glue::WebPluginDelegate {
       NPDeviceContext2D* context);
   virtual NPError Device2DSetStateContext(NPDeviceContext2D* context,
                                           int32 state,
-                                          int32 value);
+                                          intptr_t value);
   virtual NPError Device2DGetStateContext(NPDeviceContext2D* context,
                                           int32 state,
-                                          int32* value);
+                                          intptr_t* value);
   virtual NPError Device2DFlushContext(NPP id,
                                        NPDeviceContext2D* context,
                                        NPDeviceFlushContextCallbackPtr callback,
@@ -102,10 +102,10 @@ class WebPluginDelegatePepper : public webkit_glue::WebPluginDelegate {
       NPDeviceContext3D* context);
   virtual NPError Device3DSetStateContext(NPDeviceContext3D* context,
                                           int32 state,
-                                          int32 value);
+                                          intptr_t value);
   virtual NPError Device3DGetStateContext(NPDeviceContext3D* context,
                                           int32 state,
-                                          int32* value);
+                                          intptr_t* value);
   virtual NPError Device3DFlushContext(NPP id,
                                        NPDeviceContext3D* context,
                                        NPDeviceFlushContextCallbackPtr callback,
@@ -129,9 +129,9 @@ class WebPluginDelegatePepper : public webkit_glue::WebPluginDelegate {
       const NPDeviceContextAudioConfig* config,
       NPDeviceContextAudio* context);
   virtual NPError DeviceAudioSetStateContext(NPDeviceContextAudio* context,
-                                             int32 state, int32 value);
+                                             int32 state, intptr_t value);
   virtual NPError DeviceAudioGetStateContext(NPDeviceContextAudio* context,
-                                             int32 state, int32* value);
+                                             int32 state, intptr_t* value);
   virtual NPError DeviceAudioFlushContext(
       NPP id, NPDeviceContextAudio* context,
       NPDeviceFlushContextCallbackPtr callback, void* user_data);
